Share ModelForm between LLM types
